+++ src/gutenberg2zim/zim.py
@@ -4,7 +4,7 @@ import pathlib
 from peewee import fn
 
 from gutenberg2zim.constants import logger
-from gutenberg2zim.database import BookLanguage
+from gutenberg2zim.database import Book, BookLanguage
 from gutenberg2zim.export import export_all_books
 from gutenberg2zim.iso639 import ISO_MATRIX
 from gutenberg2zim.l10n import metadata_translations
@@ -13,13 +13,15 @@ from gutenberg2zim.shared import Global
 from gutenberg2zim.utils import get_project_id
 
 
-def existing_and_sorted_languages(languages: list[str] | None) -> list[str]:
+def existing_and_sorted_languages(languages: list[str] | None, books_ids) -> list[str]:
 
     # actual list of languages with books sorted by most used
     nb = fn.COUNT(BookLanguage.book).alias("nb")
     db_languages: list[str] = [
         lang.language_code
         for lang in BookLanguage.select(BookLanguage.language_code, nb)
+        .join(Book)
+        .where(not books_ids or Book.book_id << books_ids)
         .group_by(BookLanguage.language_code)
         .order_by(nb.desc())
     ]